requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
gensim==4.3.2
python-dotenv==1.0.0
matplotlib==3.8.0
wordcloud==1.9.2
openai==1.6.0